+++ apps/web/src/app/[lang]/(main)/(unirefund)/settings/templates/rebate-tables/_components/schema.ts
@@ -46,9 +46,12 @@ export function createRebateTableFormSchemas({languageData}: {languageData?: Con
 
   const createFormSchema = z
     .object({
-      name: z.string().min(2).max(50),
-      isTemplate: z.boolean().default(false),
-      calculateNetCommissionInsteadOfRefund: z.boolean().default(false),
+      name: z
+        .string()
+        .min(2, {message: languageData ? languageData["Contracts.nameIsRequired"] : ""})
+        .max(50),
+      isTemplate: z.boolean().default(false).optional(),
+      calculateNetCommissionInsteadOfRefund: z.boolean().default(false).optional(),
       merchantId: z.string().uuid().optional().nullable(),
       processingFeeDetails: processingFeeDetailsSchema,
       rebateTableDetails: rebateTableDetailsSchema,
@@ -59,7 +62,7 @@ export function createRebateTableFormSchemas({languageData}: {languageData?: Con
           ctx.addIssue({
             code: z.ZodIssueCode.custom,
             message: languageData
-              ? languageData["RebateTable.Form.rebateTableDetails.merchantSelectionIsRequiredWhenIsTemplateOptionFalse"]
+              ? languageData["RebateTable.Form.merchantSelectionIsRequiredWhenIsTemplateOptionFalse"]
               : "Merchant is required when template option is false.",
             path: ["merchantId"],
           });
@@ -68,7 +71,7 @@ export function createRebateTableFormSchemas({languageData}: {languageData?: Con
     });
   const updateFormSchema = z.object({
     name: z.string().min(2).max(50),
-    calculateNetCommissionInsteadOfRefund: z.boolean().default(false),
+    calculateNetCommissionInsteadOfRefund: z.boolean().default(false).optional(),
     processingFeeDetails: processingFeeDetailsSchema,
     rebateTableDetails: rebateTableDetailsSchema,
   });
